fix typo

+++ src/oracles/identify/service/validation.go
@@ -5,13 +5,13 @@ import (
 	"errors"
 )
 
-// 调用预言机服务的人脸比对接口返回的结果
+// 调用预言机服务的物体识别比对接口返回的结果
 type EntityIdentifyResult struct {
-	// 人脸比对结果，0表示请求成功，非0会有响应的错误码意义
-	Result int `json:"result"`
+	// 物体识别比对结果，0表示请求成功，非0会有响应的错误码意义
+	Result uint32 `json:"result"`
 
-	// 相似度评分，在0~1之间取值，越大表示越相似
-	Score float32 `json:"score"`
+	// 相似度评分，在0~100之间取值，越大表示越相似
+	Score uint32 `json:"score"`
 
 	// 会话序号
 	Sequence string `json:"sequence"`
